Submission refactor
