+++ services/madoc-ts/src/frontend/site/pages/tasks/preview-crowdsourcing-task.tsx
@@ -1,13 +1,10 @@
-import React, { Suspense, useCallback, useState } from 'react';
-import { Heading3 } from '../../../shared/atoms/Heading3';
+import React, { useState } from 'react';
 import { CrowdsourcingTask } from '../../../../types/tasks/crowdsourcing-task';
 import { useQuery } from 'react-query';
 import { useApi } from '../../../shared/hooks/use-api';
 import { ThemeProvider } from 'styled-components';
 import { defaultTheme, Revisions } from '@capture-models/editor';
 import { ViewContent } from '../../../shared/components/ViewContent';
-import { CanvasFull } from '../../../../types/schemas/canvas-full';
-import { CaptureModel, RevisionRequest } from '@capture-models/types';
 import { RevisionTopLevel } from '../../../shared/caputre-models/RevisionTopLevel';
 import {
   EditorToolbarButton,
@@ -19,355 +16,16 @@ import {
 } from '../../../shared/atoms/EditorToolbar';
 import { ArrowBackIcon } from '../../../shared/icons/ArrowBackIcon';
 import { EditIcon } from '../../../shared/icons/EditIcon';
-import { ModalButton } from '../../../shared/components/Modal';
-import { Button, ButtonRow } from '../../../shared/atoms/Button';
-import { DeleteForeverIcon } from '../../../shared/icons/DeleteForeverIcon';
-import { TextField } from '@capture-models/editor/lib/input-types/TextField/TextField';
-import { ReadMoreIcon } from '../../../shared/icons/ReadMoreIcon';
-import { CallMergeIcon } from '../../../shared/icons/CallMergeIcon';
-import { GradingIcon } from '../../../shared/icons/GradingIcon';
 import { FullScreenExitIcon } from '../../../shared/icons/FullScreenExitIcon';
 import { FullScreenEnterIcon } from '../../../shared/icons/FullScreenEnterIcon';
 import { MaximiseWindow } from '../../../shared/atoms/MaximiseWindow';
 import { PreviewIcon } from '../../../shared/icons/PreviewIcon';
-import ReactTimeago from 'react-timeago';
 import { useLoadedCaptureModel } from '../../../shared/hooks/use-loaded-capture-model';
 import { WarningMessage } from '../../../shared/atoms/WarningMessage';
-
-const RequestChanges: React.FC<{ userTaskId: string; changesRequested?: null | string; onRequest: () => void }> = ({
-  changesRequested,
-  userTaskId,
-  onRequest,
-}) => {
-  const api = useApi();
-  const [requestMessage, setRequestMessage] = useState(changesRequested || '');
-  const [isLoading, setIsLoading] = useState(false);
-  const { currentRevision } = Revisions.useStoreState(state => {
-    return {
-      currentRevision: state.currentRevision,
-    };
-  });
-  const deselectRevision = Revisions.useStoreActions(a => a.deselectRevision);
-
-  const requestChangesApiCall = useCallback(() => {
-    if (currentRevision) {
-      setIsLoading(true);
-      api
-        .reviewRequestChanges({
-          message: requestMessage,
-          revisionRequest: currentRevision,
-          userTaskId,
-        })
-        .then(() => {
-          deselectRevision({ revisionId: currentRevision.revision.id });
-          setIsLoading(false);
-          onRequest();
-        });
-    }
-  }, [api, currentRevision, deselectRevision, onRequest, requestMessage, userTaskId]);
-
-  if (currentRevision?.revision.approved) {
-    return null;
-  }
-
-  return (
-    <EditorToolbarButton
-      disabled={isLoading}
-      as={ModalButton}
-      button={true}
-      autoHeight={true}
-      title="Request changes"
-      render={() => (
-        <Suspense fallback={<div />}>
-          <label htmlFor="message">Write a message to the contributor</label>
-          <TextField
-            id="message"
-            type="text-field"
-            value={requestMessage}
-            label="Write message to the contributor"
-            updateValue={setRequestMessage}
-            multiline={true}
-          />
-          <p>Once requested the task will be assigned back to user</p>
-        </Suspense>
-      )}
-      renderFooter={({ close }: any) => (
-        <Button
-          style={{ marginLeft: 'auto' }}
-          onClick={() => {
-            close();
-            requestChangesApiCall();
-          }}
-        >
-          Request changes
-        </Button>
-      )}
-    >
-      <EditorToolbarIcon>
-        <ReadMoreIcon />
-      </EditorToolbarIcon>
-      <EditorToolbarLabel>request changes</EditorToolbarLabel>
-    </EditorToolbarButton>
-  );
-};
-
-const ApproveSubmission: React.FC<{
-  onApprove: () => void;
-  userTaskId: string;
-  allUserTaskIds: string[];
-  allRevisionIds: string[];
-  reviewTaskId: string;
-}> = ({ userTaskId, allUserTaskIds, allRevisionIds, reviewTaskId, onApprove }) => {
-  const { acceptedRevision } = Revisions.useStoreState(state => {
-    return {
-      acceptedRevision: state.currentRevision,
-    };
-  });
-  const [loading, setIsLoading] = useState(false);
-  const api = useApi();
-  const revisionIdsToRemove = allRevisionIds.filter(id => id && id !== acceptedRevision?.revision.id);
-  const userTaskIdsToRemove = allUserTaskIds.filter(id => id && id !== userTaskId);
-  const approveAndRemoveApiCall = useCallback(() => {
-    if (acceptedRevision) {
-      setIsLoading(true);
-      api
-        .reviewApproveAndRemoveSubmission({
-          userTaskIds: userTaskIdsToRemove,
-          reviewTaskId,
-          acceptedRevision,
-          revisionIdsToRemove,
-        })
-        .then(() => {
-          setIsLoading(false);
-          onApprove();
-        });
-    }
-  }, [acceptedRevision, api, onApprove, reviewTaskId, revisionIdsToRemove, userTaskIdsToRemove]);
-
-  const approveApiCall = useCallback(() => {
-    if (acceptedRevision) {
-      setIsLoading(true);
-      api
-        .reviewApproveSubmission({
-          revisionRequest: acceptedRevision,
-          userTaskId,
-        })
-        .then(() => {
-          setIsLoading(false);
-          onApprove();
-        });
-    }
-  }, [acceptedRevision, api, onApprove, userTaskId]);
-
-  if (acceptedRevision?.revision.status === 'accepted') {
-    return null;
-  }
-
-  return (
-    <EditorToolbarButton
-      as={ModalButton}
-      button={true}
-      disabled={loading}
-      autoHeight={true}
-      title="Approve submission"
-      render={() => (
-        <div>
-          <ul>
-            <li>
-              <strong>Approve</strong> - The submission will be approved and all other submission will remain
-            </li>
-            <li>
-              <strong>Approve and remove remaining</strong> - The submission will be approved and all other submission
-              will be removed. The users who created all of the submissions will see their submission approved. This can
-              be good if merging multiple revisions.
-            </li>
-          </ul>
-        </div>
-      )}
-      renderFooter={({ close }: any) => (
-        <ButtonRow style={{ margin: '0 0 0 auto' }}>
-          <Button
-            onClick={() => {
-              approveApiCall();
-              close();
-            }}
-          >
-            Approve
-          </Button>
-          <Button
-            onClick={() => {
-              approveAndRemoveApiCall();
-              close();
-            }}
-          >
-            Approve and remove remaining
-          </Button>
-        </ButtonRow>
-      )}
-    >
-      <EditorToolbarIcon>
-        <GradingIcon />
-      </EditorToolbarIcon>
-      <EditorToolbarLabel>approve</EditorToolbarLabel>
-    </EditorToolbarButton>
-  );
-};
-
-const RejectSubmission: React.FC<{ onReject: () => void; userTaskId: string }> = ({ onReject, userTaskId }) => {
-  const api = useApi();
-  const [isLoading, setIsLoading] = useState(false);
-  const { currentRevision } = Revisions.useStoreState(state => {
-    return {
-      currentRevision: state.currentRevision,
-    };
-  });
-  const deselectRevision = Revisions.useStoreActions(a => a.deselectRevision);
-
-  const rejectApiCall = useCallback(() => {
-    if (currentRevision) {
-      setIsLoading(true);
-      api
-        .reviewRejectSubmission({
-          revisionRequest: currentRevision,
-          userTaskId,
-        })
-        .then(() => {
-          deselectRevision({ revisionId: currentRevision.revision.id });
-          setIsLoading(false);
-          onReject();
-        });
-    }
-  }, [api, currentRevision, deselectRevision, onReject, userTaskId]);
-
-  if (!currentRevision || currentRevision.revision.approved) {
-    return null;
-  }
-
-  return (
-    <EditorToolbarButton
-      as={ModalButton}
-      disabled={isLoading}
-      button={true}
-      autoHeight={true}
-      title="Reject submission"
-      render={() => (
-        <div>
-          <strong>Are you sure you want to delete this revision and mark the task as rejected?</strong>
-          <ul>
-            <li>The user will be notified that the revision has been rejected</li>
-            <li>You will no longer be able to see the content in the revision</li>
-          </ul>
-        </div>
-      )}
-      renderFooter={({ close }: any) => (
-        <Button
-          style={{ marginLeft: 'auto' }}
-          onClick={() => {
-            close();
-            rejectApiCall();
-          }}
-        >
-          Reject changes
-        </Button>
-      )}
-    >
-      <EditorToolbarIcon>
-        <DeleteForeverIcon />
-      </EditorToolbarIcon>
-      <EditorToolbarLabel>reject submission</EditorToolbarLabel>
-    </EditorToolbarButton>
-  );
-};
-
-const StartMerge: React.FC<{
-  userTask: CrowdsourcingTask & { id: string };
-  allTasks: Array<CrowdsourcingTask & { id: string }>;
-  reviewTaskId: string;
-  onStartMerge: (taskId: string) => void;
-}> = ({ reviewTaskId, onStartMerge, allTasks, userTask }) => {
-  const [selected, setSelected] = useState(() => allTasks.filter(t => t.id !== userTask.id).map(t => t.id as string));
-  const [isLoading, setIsLoading] = useState(false);
-  const { currentRevision } = Revisions.useStoreState(state => {
-    return {
-      currentRevision: state.currentRevision,
-    };
-  });
-  const deselectRevision = Revisions.useStoreActions(a => a.deselectRevision);
-  const api = useApi();
-
-  const startMergeApiCall = useCallback(() => {
-    if (currentRevision) {
-      setIsLoading(true);
-      api
-        .reviewPrepareMerge({
-          reviewTaskId: reviewTaskId,
-          revision: currentRevision,
-          toMerge: selected,
-          revisionTask: userTask.id,
-        })
-        .then(req => {
-          deselectRevision({ revisionId: currentRevision.revision.id });
-          setIsLoading(false);
-          onStartMerge(req.revision.id);
-        });
-    }
-  }, [api, currentRevision, deselectRevision, onStartMerge, reviewTaskId, selected]);
-
-  return (
-    <EditorToolbarButton
-      as={ModalButton}
-      button={true}
-      autoHeight={true}
-      title="Prepare merge"
-      render={() => (
-        <div>
-          <div>Merge the following submissions into the base</div>
-          <ul>
-            {allTasks.map(task =>
-              userTask.id === task.id ? null : (
-                <li key={task.id}>
-                  <label>
-                    <input
-                      type="checkbox"
-                      checked={selected.indexOf(task.id) !== -1}
-                      onChange={() => {
-                        setSelected(list => {
-                          if (list.indexOf(task.id) === -1) {
-                            return [...list, task.id];
-                          }
-                          return list.filter(t => t !== task.id);
-                        });
-                      }}
-                    />
-                    {task.creator?.name || task.name}{' '}
-                    {task.modified_at ? <ReactTimeago date={task.modified_at} /> : null}
-                  </label>
-                </li>
-              )
-            )}
-          </ul>
-        </div>
-      )}
-      renderFooter={({ close }: any) => (
-        <Button
-          style={{ marginLeft: 'auto' }}
-          disabled={selected.length === 0 || isLoading}
-          onClick={() => {
-            startMergeApiCall();
-            close();
-          }}
-        >
-          Start merge
-        </Button>
-      )}
-    >
-      <EditorToolbarIcon>
-        <CallMergeIcon />
-      </EditorToolbarIcon>
-      <EditorToolbarLabel>start merge</EditorToolbarLabel>
-    </EditorToolbarButton>
-  );
-};
+import { RequestChanges } from './actions/request-changes';
+import { ApproveSubmission } from './actions/approve-submission';
+import { RejectSubmission } from './actions/reject-submission';
+import { StartMerge } from './actions/start-merge';
 
 const PreviewCrowdsourcingTask: React.FC<{
   task: CrowdsourcingTask & { id: string };
@@ -392,6 +50,7 @@ const PreviewCrowdsourcingTask: React.FC<{
   const [{ captureModel, canvas }] = useLoadedCaptureModel(modelId);
 
   const isLocked = props.lockedTasks && props.lockedTasks.indexOf(props.task.id) !== -1;
+  const isDone = taskData?.status === 3;
 
   return (
     <ThemeProvider theme={defaultTheme}>
@@ -410,12 +69,12 @@ const PreviewCrowdsourcingTask: React.FC<{
 
                 <EditorToolbarSpacer />
 
-                <EditorToolbarButton onClick={() => setIsEditing(r => !r)} disabled={isLocked}>
+                <EditorToolbarButton onClick={() => setIsEditing(r => !r)} disabled={isLocked || isDone}>
                   <EditorToolbarIcon>{isEditing ? <PreviewIcon /> : <EditIcon />}</EditorToolbarIcon>
                   <EditorToolbarLabel>{isEditing ? 'preview submission' : 'edit submission'}</EditorToolbarLabel>
                 </EditorToolbarButton>
 
-                {isLocked ? null : (
+                {isLocked || isDone ? null : (
                   <>
                     <RejectSubmission userTaskId={props.task.id} onReject={() => props.goBack({ refresh: true })} />
 
@@ -453,6 +112,7 @@ const PreviewCrowdsourcingTask: React.FC<{
                 <div style={{ width: '33%', padding: '1em' }}>
                   <RevisionTopLevel
                     allowEdits={false}
+                    allowNavigation={false}
                     onSaveRevision={async rev => {
                       console.log(rev);
                     }}
